packages/model/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aistudio-sdk==0.3.8
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.11.0
anytree==2.13.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.4.0
bce-python-sdk==0.9.46
beautifulsoup4==4.14.2
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.6
decorator==5.2.1
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.119.1
ffmpy==0.6.4
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gradio==5.4.0
gradio_client==1.4.2
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.0
idna==3.11
imageio==2.37.0
imagesize==1.4.1
imbalanced-learn==0.13.0
imgaug==0.4.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.15.2
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
modelscope==1.31.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
openai==2.6.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
packaging==24.0
paddleocr==2.9.1
paddlepaddle-gpu==2.6.1
paddlex==3.3.5
pandas==2.2.3
pillow==11.3.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pymilvus==2.6.2
pyparsing==3.2.5
pypdfium2==4.30.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.12
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
ruff==0.14.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.4
-e git+https://github.com/hshlalla/Smart_vision.git@82c10a6062d2a17bfa1ced187bb00cb8141f2603#egg=smart_match_model&subdirectory=packages/model
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tokenizers==0.21.4
tomli==2.3.0
tomlkit==0.12.0
torch==2.7.1
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.51.3
triton==3.3.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==12.0
yarl==1.22.0
